Implement RSC-First Hybrid Rendering
