+++ sdk/src/runtime/render/renderToStream.tsx
@@ -42,7 +42,6 @@ export const renderToStream = async (
 
   const htmlStream = await transformRscToHtmlStream({
     stream: rscStream,
-    Document,
     requestInfo,
     onError,
   });
