+++ sdk/src/runtime/render/transformRscToHtmlStream.tsx
@@ -1,6 +1,5 @@
 import { createModuleMap } from "./createModuleMap.js";
 import ReactServerDom from "react-server-dom-webpack/client.edge";
-import { DocumentProps } from "../lib/router";
 import { renderRscThenableToHtmlStream } from "rwsdk/__ssr_bridge";
 import { RequestInfo } from "../requestInfo/types";
 
@@ -8,12 +7,10 @@ const { createFromReadableStream } = ReactServerDom;
 
 export const transformRscToHtmlStream = ({
   stream,
-  Document,
   requestInfo,
   onError,
 }: {
   stream: ReadableStream;
-  Document: React.FC<DocumentProps>;
   requestInfo: RequestInfo;
   onError: (error: unknown) => void;
 }) => {
@@ -26,9 +23,7 @@ export const transformRscToHtmlStream = ({
 
   return renderRscThenableToHtmlStream({
     thenable,
-    Document,
     requestInfo,
-    shouldSSR: requestInfo.rw.ssr,
     onError,
   });
 };
